Try fix restart
Cancel event listener after 5 wins

Still trying to find a way around the fact that is I use a while
loop for keepPlaying it just breaks the browser. Maybe if I could
prompt for it to wait a certain event to happen but it is still very
much obscure and out of the scop from what I have learnt so far. Either
way I'm making porgress though the code has been very confusing for
anyone who tries to read it.